requirement.txt
django
djangorestframework
pydub
pyaudio
mysql-python
mysqlclient
python-docx
numpy
baidu-aip
